Only update port if changed

+++ start.sh
@@ -1,14 +1,18 @@
 #!/bin/bash
 
 COOKIES="/tmp/cookies.txt"
+CURRENT_PORT=""
 
 update_port () {
-  PORT=$(cat $PORT_FORWARDED)
-  rm -f $COOKIES
-  curl -s -c $COOKIES --data "username=$QBITTORRENT_USER&password=$QBITTORRENT_PASS" ${HTTP_S}://${QBITTORRENT_SERVER}:${QBITTORRENT_PORT}/api/v2/auth/login > /dev/null
-  curl -s -b $COOKIES --data 'json={"listen_port": "'"$PORT"'"}' ${HTTP_S}://${QBITTORRENT_SERVER}:${QBITTORRENT_PORT}/api/v2/app/setPreferences > /dev/null
-  rm -f $COOKIES
-  echo "Successfully updated qbittorrent to port $PORT"
+  NEW_PORT=$(cat $PORT_FORWARDED)
+  if [ "$NEW_PORT" != "$CURRENT_PORT" ]; then
+    CURRENT_PORT=$NEW_PORT
+    rm -f $COOKIES
+    curl -s -c $COOKIES --data "username=$QBITTORRENT_USER&password=$QBITTORRENT_PASS" ${HTTP_S}://${QBITTORRENT_SERVER}:${QBITTORRENT_PORT}/api/v2/auth/login > /dev/null
+    curl -s -b $COOKIES --data 'json={"listen_port": "'"$CURRENT_PORT"'"}' ${HTTP_S}://${QBITTORRENT_SERVER}:${QBITTORRENT_PORT}/api/v2/app/setPreferences > /dev/null
+    rm -f $COOKIES
+    echo "Successfully updated qbittorrent to port $CURRENT_PORT"
+  fi
 }
 
 while true; do
